🚚 Rename articles directory --- This allows us to drop the "slug" field that we create on the node to replace "articles" with "blog"
