+++ src/content/blog/de/laravel-package.mdx
