+++ src/content/blog/de/laravel-repository-pattern.mdx
