+++ src/content/blog/de/vue-app-auth.mdx
