+++ src/content/blog/de/vuex-einstieg.mdx
