+++ src/content/blog/en/closing-maxout.mdx
