+++ src/content/blog/en/from-ghent-with-love.mdx
