+++ src/content/blog/en/how-to-slack.mdx
