+++ src/content/blog/en/review-2018.mdx
