+++ src/content/blog/en/review-2019.mdx
