+++ src/content/blog/externalLinks.ts
